refactor:將購物車更新程式碼包到service
