+++ app/Services/CartItemStoreService.php
@@ -6,7 +6,7 @@ use App\Models\Race;
 use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
-class CartItemService
+class CartItemStoreService
 {
     protected $userId;
     protected $quantity;
@@ -46,43 +46,5 @@ class CartItemService
         $this->cartItem->quantity = $newQuantity;
         $this->cartItem->save();
         return ['success' => config('success_messages.ITEM_ADD_TO_CART'), 'status' => Response::HTTP_OK];
-
-
-namespace App\Services;
-
-use App\Models\CartItem;
-
-class CartItemService
-{
-    public function updateCartItemAndCalculateTotal($userId, $raceId, $quantity)
-    {
-        $this->updateCartItem($userId, $raceId, $quantity);
-        return $this->calculateTotalPrice($userId);
-    }
-
-    private function updateCartItem($userId, $raceId, $quantity)
-    {
-        $cartItem = CartItem::where('user_id', $userId)->where('race_id', $raceId)->first();
-
-        if (!$cartItem) {
-            // 處理找不到 CartItem 的情況
-            throw new \Exception(config("error_messages.NO_CART_ITEM"));
-        }
-
-        $cartItem->update(['quantity' => $quantity]);
-    }
-
-    private function calculateTotalPrice($userId)
-    {
-        $total = CartItem::where('user_id', $userId)
-            ->selectRaw('SUM(current_price * quantity) as total')
-            ->value('total');
-
-        if ($total === null) {
-            // 處理無法計算總金額的情況
-            throw new \Exception(config("error_messages.CACULATION_FAILED"));
-        }
-
-        return $total;
     }
 }
