+++ config/error_messages.php
@@ -4,12 +4,13 @@
 return [
     'EMAIL_NOT_VERIFIED' => 'email not verify',
     'INVALID_CREDENTIALS' => 'Invalid credentials',
-    'INVALID_CREDENTIALS' => 'Invalid credentials',
     'LOGOUT_FAILED' => 'Failed to logout',
     'REDIRECT_FAILED' => 'Failed to redirect',
     "EMAIL_HAS_REGISTERED"=>'Email already, registered.',
      "QUANTITY_EXCEED_STOCK"=>'Requested quantity exceeds available stock',
-     "QUANTITY_EXCEED_STOCK"=>'Requested quantity exceeds available stock',
+    "NO_CART_ITEM"=>"No cart item found for the given user and race.",
+    "CACULATION_FAILED"=>"Unable to calculate total price.",
+
 
     
     // ... 其他錯誤訊息
